feat: Add and modularize record existence check functions, test: Implement API-specific test codes

+++ src/services/like.service.js
@@ -1,23 +1,36 @@
 const { HttpError } = require('../middlewares/errorHandler.middleware');
 const { StatusCodes } = require('http-status-codes');
+const { isLikeExist } = require('../utils/existanceCheck.util');
 
 const addToLikes = async (conn, userId, bookId) => {
-    const sql = "INSERT INTO likes (user_id, book_id) VALUES (?, ?)";
-    const values = [userId, bookId];
-    const [result] = await conn.query(sql, values);
+    const exists = await isLikeExist(conn, userId, bookId);
 
-    return result;
+    if (!exists) {
+        const sql = "INSERT INTO likes (user_id, book_id) VALUES (?, ?)";
+        const values = [userId, bookId];
+        const [result] = await conn.query(sql, values);
+
+        return result;
+    } else {
+        throw new HttpError(StatusCodes.CONFLICT, "이미 처리된 요청입니다.");
+    }
 };
 
 const removeFromLikes = async (conn, userId, bookId) => {
-    const sql = "DELETE FROM likes WHERE user_id = ? AND book_id = ?";
-    const values = [userId, bookId];
-    const [result] = await conn.query(sql, values);
+    const exists = await isLikeExist(conn, userId, bookId);
 
-    if (result.affectedRows) {
-        return result;
+    if (exists) {
+        const sql = "DELETE FROM likes WHERE user_id = ? AND book_id = ?";
+        const values = [userId, bookId];
+        const [result] = await conn.query(sql, values);
+
+        if (result.affectedRows) {
+            return result;
+        } else {
+            throw new HttpError(StatusCodes.BAD_REQUEST);
+        }
     } else {
-        throw new HttpError(StatusCodes.BAD_REQUEST);
+        throw new HttpError(StatusCodes.CONFLICT, "이미 처리된 요청입니다.");
     }
 };
